+++ Tests/Shared.Specs/Collections/GenericCollectionAssertionOfStringSpecs.cs
@@ -1702,7 +1702,7 @@ namespace FluentAssertions.Specs
 
             // Assert
             action.Should().Throw<XunitException>()
-                .WithMessage("Expected collection {\"build succeded\", \"test failed\"} to not contain a match of \"* failed\" because it shouldn't.");
+                .WithMessage("Did not expect collection {\"build succeded\", \"test failed\"} to contain a match of \"* failed\" because it shouldn't.");
         }
 
         [Fact]
@@ -1716,7 +1716,7 @@ namespace FluentAssertions.Specs
 
             // Assert
             action.Should().Throw<XunitException>()
-                .WithMessage("Expected collection {\"build failed\", \"test failed\"} to not contain a match of \"* failed\" because it shouldn't.");
+                .WithMessage("Did not expect collection {\"build failed\", \"test failed\"} to contain a match of \"* failed\" because it shouldn't.");
         }
 
         [Fact]
@@ -1744,7 +1744,26 @@ namespace FluentAssertions.Specs
 
             // Assert
             action.Should().Throw<XunitException>()
-                .WithMessage("Expected collection to not contain a match of <null>, but found <null>.");
+                .WithMessage("Did not expect collection to contain a match of <null>, but found <null>.");
+        }
+
+        [Fact]
+        public void When_collection_contains_multiple_matches_for_not_match_which_should_throw()
+        {
+            // Arrange
+            IEnumerable<string> collection = new string[] { "build succeded", "test failed", "pack failed" };
+
+            // Act
+            Action action = () =>
+            {
+                string item = collection.Should().NotContainMatch("* failed").Which;
+            };
+
+            // Assert
+            action.Should().Throw<XunitException>()
+               .WithMessage("Did not expect string item = collection {\"build succeded\", \"test failed\", \"pack failed\"} to contain a match of \"* failed\".");
+
+
         }